This resolves #2810
I made an attempt to implement it without using the new field.
The problem I ran into was both parseJob and parseGroup calling
parseUpdate, and to calculate a percentage of count, you need the
count. This means you will need a new arguement to parseUpdate, and
calling with an empty string from parseJob.
I clean up the code I wrote for this if the above approach is preferred.