Remove logging and add more error handling
Remove logging used for debugging.
Better error handling, e.g if new group count < max parallel or 0,
return an error. DO NOT write the job as these values are saved in the
job and later returned in as an error which is hard to parse (and get out of).
If no task group was modified, do not write a job and return the job
file