meta: Set main.VersionPrerelease="dev"
For most projects using this versioning system, making a commit bumping
the version and re-adding the development tag is part of the release
SOP (certainly it was for Terraform!). I'm not sure why it didn't happen
here, but without this, all development builds show up as Nomad v0.6.0
with no dirty marker or commit marker.