Handle exact namespace match properly when also matched as prefix
This fixes an issue with the `nomad namespace inspect` and `nomad namespace status` commands failing to run for namespaces whose names are the start of another namespace's name: for example, foo and foobar.  If the argument matches the first possibility returned by prefix exactly, that namespace will be targetted.  If no exact match occurs, the behavior is unchanged.