Made only one return value non-nil; fixed logic issue w/ sort
Per @schmichael, better to have just one of the return values non-nil.
Also corrected issue with test lying because the return was ordered by
creation time.  Added an additional namespace to prevent ordering
accidentally make the test pass and verified that the proper namespace
was being returned with a string test on the output.