Update template.html.md Vault template example
The change updates the Vault example within the Environment Variables section to use the `.value` notation rather than the `.key` notation. This is because the example is calling a specific secret key and populating an env var with this; meaning the value of the secret is required over the value.