Implementation of job restart logic
This feature is useful because there are instances where you wish to
restart the existing job definition, triggering any update/canary logic
that exists.

There are two ways to restart a job:

`nomad run -restart jobfile.nomad`

If the job already exists, a restart/update will be triggered even if
the job definition is identical to the previous version.  If the job
does not exist it will be registered and executed.

`nomad restart jobid`

This will restart an existing job.  This creates a new job version and
triggers the update/canary logic.