Set an API http client read timeout through the deadline
By setting the transport timeout we are only applying the timeout to the
connection establishment. By adding 1 line in which we set the timeout
of the http client we are effectively using a context for each request
which in turns allows us to set a read timeout.

I had to do this because some calls, under high concurrency, tend to lock
forever.