Report driver schema errors in `alloc status`
Currently, if a driver returns a bad schema, the error is logged on
client but is never showed in `nomad alloc status ...`.

With this change, we report conversion driver schema diagnostics errors to user the same way as if user passed a bad config file.

The advantage in our context is that interpolation errors from driver
schema defaults get reported to users.