drivers: Handle out-of-nomad executor termination
This commit adds a new handleFailedExecutor function to drivers that
execute processes via the nomad `executor` supervisor.

This function is invoked during failures in WaitTask to avoid the case where
we will orphan the users application process if the executor process is
killed unexpectedly, as the child may continue to live on.

It does _not_ fix the leak in cases where an executor process dies while the
Task Driver process is down, because in those cases we will need to more
strictly validate the child pid to avoid wrongfully terminating user
processes in high pid churn environments, or after extended client
downtime.